Refa: restore openai-compatible chat completions api
