+++ test/testcases/test_http_api/test_session_management/test_session_sdk_routes_unit.py
@@ -667,6 +667,34 @@ def _load_agent_api_module(monkeypatch):
     return module
 
 
+def _load_openai_api_module(monkeypatch):
+    _load_session_module(monkeypatch)
+    repo_root = Path(__file__).resolve().parents[4]
+
+    api_apps_mod = ModuleType("api.apps")
+    api_apps_mod.__path__ = [str(repo_root / "api" / "apps")]
+    api_apps_mod.login_required = lambda func: func
+    api_apps_mod.current_user = SimpleNamespace(id="tenant-1")
+    monkeypatch.setitem(sys.modules, "api.apps", api_apps_mod)
+
+    api_apps_restful_mod = ModuleType("api.apps.restful_apis")
+    api_apps_restful_mod.__path__ = [str(repo_root / "api" / "apps" / "restful_apis")]
+    monkeypatch.setitem(sys.modules, "api.apps.restful_apis", api_apps_restful_mod)
+
+    quart_mod = ModuleType("quart")
+    quart_mod.Response = _StubResponse
+    quart_mod.jsonify = lambda payload: payload
+    monkeypatch.setitem(sys.modules, "quart", quart_mod)
+
+    module_path = repo_root / "api" / "apps" / "restful_apis" / "openai_api.py"
+    spec = importlib.util.spec_from_file_location("test_openai_api_unit_module", module_path)
+    module = importlib.util.module_from_spec(spec)
+    module.manager = _DummyManager()
+    monkeypatch.setitem(sys.modules, "test_openai_api_unit_module", module)
+    spec.loader.exec_module(module)
+    return module
+
+
 @pytest.mark.p2
 def test_create_and_update_guard_matrix(monkeypatch):
     module = _load_session_module(monkeypatch)
@@ -739,10 +767,14 @@ def test_chat_completion_metadata_and_stream_paths(monkeypatch):
 
 @pytest.mark.p2
 def test_openai_chat_validation_matrix_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
+    module = _load_openai_api_module(monkeypatch)
 
     monkeypatch.setattr(module, "num_tokens_from_string", lambda _text: 1)
-    monkeypatch.setattr(module.DialogService, "query", lambda **_kwargs: [SimpleNamespace(kb_ids=["kb-1"])])
+    monkeypatch.setattr(
+        module.DialogService,
+        "query",
+        lambda **_kwargs: [SimpleNamespace(kb_ids=["kb-1"], llm_id="chat-model", tenant_id="tenant-1")],
+    )
 
     cases = [
         (
@@ -786,20 +818,23 @@ def test_openai_chat_validation_matrix_unit(monkeypatch):
 
     for payload, expected in cases:
         monkeypatch.setattr(module, "get_request_json", lambda p=payload: _AwaitableValue(p))
-        res = _run(inspect.unwrap(module.chat_completion_openai_like)("tenant-1", "chat-1"))
+        res = _run(inspect.unwrap(module.openai_chat_completions)("chat-1"))
         assert expected in res["message"]
 
 
 @pytest.mark.p2
 def test_openai_stream_generator_branches_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
+    module = _load_openai_api_module(monkeypatch)
 
-    monkeypatch.setattr(module, "Response", _StubResponse)
     monkeypatch.setattr(module, "num_tokens_from_string", lambda text: len(text or ""))
     monkeypatch.setattr(module, "convert_conditions", lambda cond: cond.get("conditions", []))
     monkeypatch.setattr(module, "meta_filter", lambda *_args, **_kwargs: [])
     monkeypatch.setattr(module.DocMetadataService, "get_flatted_meta_by_kbs", lambda _kb_ids: [{"id": "doc-1"}])
-    monkeypatch.setattr(module.DialogService, "query", lambda **_kwargs: [SimpleNamespace(kb_ids=["kb-1"])])
+    monkeypatch.setattr(
+        module.DialogService,
+        "query",
+        lambda **_kwargs: [SimpleNamespace(kb_ids=["kb-1"], llm_id="chat-model", tenant_id="tenant-1")],
+    )
     monkeypatch.setattr(module, "_build_reference_chunks", lambda *_args, **_kwargs: [{"id": "ref-1"}])
 
     async def fake_async_chat(_dia, _msg, _stream, **_kwargs):
@@ -829,7 +864,7 @@ def test_openai_stream_generator_branches_unit(monkeypatch):
     }
     monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue(payload))
 
-    resp = _run(inspect.unwrap(module.chat_completion_openai_like)("tenant-1", "chat-1"))
+    resp = _run(inspect.unwrap(module.openai_chat_completions)("chat-1"))
     assert isinstance(resp, _StubResponse)
     assert resp.headers.get("Content-Type") == "text/event-stream; charset=utf-8"
 
@@ -843,11 +878,14 @@ def test_openai_stream_generator_branches_unit(monkeypatch):
 
 @pytest.mark.p2
 def test_openai_nonstream_branch_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
+    module = _load_openai_api_module(monkeypatch)
 
-    monkeypatch.setattr(module, "jsonify", lambda payload: payload)
     monkeypatch.setattr(module, "num_tokens_from_string", lambda text: len(text or ""))
-    monkeypatch.setattr(module.DialogService, "query", lambda **_kwargs: [SimpleNamespace(kb_ids=[])])
+    monkeypatch.setattr(
+        module.DialogService,
+        "query",
+        lambda **_kwargs: [SimpleNamespace(kb_ids=[], llm_id="chat-model", tenant_id="tenant-1")],
+    )
 
     async def fake_async_chat(_dia, _msg, _stream, **_kwargs):
         yield {"answer": "world", "reference": {}}
@@ -865,7 +903,7 @@ def test_openai_nonstream_branch_unit(monkeypatch):
         ),
     )
 
-    res = _run(inspect.unwrap(module.chat_completion_openai_like)("tenant-1", "chat-1"))
+    res = _run(inspect.unwrap(module.openai_chat_completions)("chat-1"))
     assert res["choices"][0]["message"]["content"] == "world"
     
 
@@ -1113,92 +1151,6 @@ def test_delete_agent_session_error_matrix_unit(monkeypatch):
     res = _run(inspect.unwrap(module.delete_agent_session)("tenant-1", "agent-1"))
     assert res["data"]["success_count"] == 1
     assert res["data"]["errors"] == ["Duplicate session ids: ok"]
-
-
-@pytest.mark.p2
-def test_sessions_ask_route_validation_and_stream_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
-    monkeypatch.setattr(module, "Response", _StubResponse)
-
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"dataset_ids": ["kb-1"]}))
-    res = _run(inspect.unwrap(module.ask_about)("tenant-1"))
-    assert res["message"] == "`question` is required."
-
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"question": "q"}))
-    res = _run(inspect.unwrap(module.ask_about)("tenant-1"))
-    assert res["message"] == "`dataset_ids` is required."
-
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"question": "q", "dataset_ids": "kb-1"}))
-    res = _run(inspect.unwrap(module.ask_about)("tenant-1"))
-    assert res["message"] == "`dataset_ids` should be a list."
-
-    monkeypatch.setattr(module.KnowledgebaseService, "accessible", lambda *_args, **_kwargs: False)
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"question": "q", "dataset_ids": ["kb-1"]}))
-    res = _run(inspect.unwrap(module.ask_about)("tenant-1"))
-    assert res["message"] == "You don't own the dataset kb-1."
-
-    monkeypatch.setattr(module.KnowledgebaseService, "accessible", lambda *_args, **_kwargs: True)
-    monkeypatch.setattr(module.KnowledgebaseService, "query", lambda **_kwargs: [SimpleNamespace(chunk_num=0)])
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"question": "q", "dataset_ids": ["kb-1"]}))
-    res = _run(inspect.unwrap(module.ask_about)("tenant-1"))
-    assert res["message"] == "The dataset kb-1 doesn't own parsed file"
-
-    monkeypatch.setattr(module.KnowledgebaseService, "query", lambda **_kwargs: [SimpleNamespace(chunk_num=1)])
-    captured = {}
-
-    async def _streaming_async_ask(question, kb_ids, uid):
-        captured["question"] = question
-        captured["kb_ids"] = kb_ids
-        captured["uid"] = uid
-        yield {"answer": "first"}
-        raise RuntimeError("ask stream boom")
-
-    monkeypatch.setattr(module, "async_ask", _streaming_async_ask)
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"question": "q", "dataset_ids": ["kb-1"]}))
-    resp = _run(inspect.unwrap(module.ask_about)("tenant-1"))
-    assert isinstance(resp, _StubResponse)
-    assert resp.headers.get("Content-Type") == "text/event-stream; charset=utf-8"
-    chunks = _run(_collect_stream(resp.body))
-    assert any('"answer": "first"' in chunk for chunk in chunks)
-    assert any('"code": 500' in chunk and "**ERROR**: ask stream boom" in chunk for chunk in chunks)
-    assert '"data": true' in chunks[-1].lower()
-    assert captured == {"question": "q", "kb_ids": ["kb-1"], "uid": "tenant-1"}
-
-
-@pytest.mark.p2
-def test_sessions_related_questions_prompt_build_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
-
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({}))
-    res = _run(inspect.unwrap(module.related_questions)("tenant-1"))
-    assert res["message"] == "`question` is required."
-
-    captured = {}
-
-    class _FakeLLMBundle:
-        def __init__(self, *args, **kwargs):
-            captured["bundle_args"] = args
-            captured["bundle_kwargs"] = kwargs
-
-        async def async_chat(self, prompt, messages, options):
-            captured["prompt"] = prompt
-            captured["messages"] = messages
-            captured["options"] = options
-            return "1. First related\n2. Second related\nplain text"
-
-    monkeypatch.setattr(module, "LLMBundle", _FakeLLMBundle)
-    monkeypatch.setattr(
-        module,
-        "get_request_json",
-        lambda: _AwaitableValue({"question": "solar energy", "industry": "renewables"}),
-    )
-    res = _run(inspect.unwrap(module.related_questions)("tenant-1"))
-    assert res["data"] == ["First related", "Second related"]
-    assert "Keep the term length between 2-4 words" in captured["prompt"]
-    assert "related terms can also help search engines" in captured["prompt"]
-    assert "Ensure all search terms are relevant to the industry: renewables." in captured["prompt"]
-    assert "Keywords: solar energy" in captured["messages"][0]["content"]
-    assert captured["options"] == {"temperature": 0.9}
 
 
 @pytest.mark.p2
@@ -1702,132 +1654,8 @@ def test_searchbots_mindmap_embedded_matrix_unit(monkeypatch):
 
 
 @pytest.mark.p2
-def test_sequence2txt_embedded_validation_and_stream_matrix_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
-    handler = inspect.unwrap(module.sequence2txt)
-    monkeypatch.setattr(module, "Response", _StubResponse)
-    monkeypatch.setattr(module.tempfile, "mkstemp", lambda suffix: (11, f"/tmp/audio{suffix}"))
-    monkeypatch.setattr(module.os, "close", lambda _fd: None)
-
-    def _set_request(form, files):
-        monkeypatch.setattr(
-            module,
-            "request",
-            SimpleNamespace(form=_AwaitableValue(form), files=_AwaitableValue(files)),
-        )
-
-    _set_request({"stream": "false"}, {})
-    res = _run(handler("tenant-1"))
-    assert "Missing 'file' in multipart form-data" in res["message"]
-
-    _set_request({"stream": "false"}, {"file": _DummyUploadFile("bad.txt")})
-    res = _run(handler("tenant-1"))
-    assert "Unsupported audio format: .txt" in res["message"]
-
-    _set_request({"stream": "false"}, {"file": _DummyUploadFile("audio.wav")})
-    tenant_llm_service = sys.modules["api.db.services.tenant_llm_service"]
-    monkeypatch.setattr(tenant_llm_service.TenantService, "get_by_id", lambda _tid: (False, None))
-    res = _run(handler("tenant-1"))
-    assert res["message"] == "Tenant not found!"
-
-    _set_request({"stream": "false"}, {"file": _DummyUploadFile("audio.wav")})
-    tenant_llm_service = sys.modules["api.db.services.tenant_llm_service"]
-    monkeypatch.setattr(tenant_llm_service.TenantService, "get_by_id", lambda _tid: (True, SimpleNamespace(asr_id="", tts_id="", llm_id="", embd_id="", img2txt_id="", rerank_id="")))
-    res = _run(handler("tenant-1"))
-    assert res["message"] == "No default ASR model is set"
-
-    class _SyncASR:
-        def transcription(self, _path):
-            return "transcribed text"
-
-        def stream_transcription(self, _path):
-            return []
-
-    _set_request({"stream": "false"}, {"file": _DummyUploadFile("audio.wav")})
-    monkeypatch.setattr(tenant_llm_service.TenantService, "get_by_id", lambda _tid: (True, SimpleNamespace(asr_id="asr-x", tts_id="", llm_id="", embd_id="", img2txt_id="", rerank_id="")))
-    monkeypatch.setattr(module, "LLMBundle", lambda *_args, **_kwargs: _SyncASR())
-    monkeypatch.setattr(module.os, "remove", lambda _path: (_ for _ in ()).throw(RuntimeError("cleanup fail")))
-    res = _run(handler("tenant-1"))
-    assert res["code"] == 0
-    assert res["data"]["text"] == "transcribed text"
-
-    class _StreamASR:
-        def transcription(self, _path):
-            return ""
-
-        def stream_transcription(self, _path):
-            yield {"event": "partial", "text": "hello"}
-
-    _set_request({"stream": "true"}, {"file": _DummyUploadFile("audio.wav")})
-    monkeypatch.setattr(module, "LLMBundle", lambda *_args, **_kwargs: _StreamASR())
-    monkeypatch.setattr(module.os, "remove", lambda _path: None)
-    resp = _run(handler("tenant-1"))
-    assert isinstance(resp, _StubResponse)
-    assert resp.content_type == "text/event-stream"
-    chunks = _run(_collect_stream(resp.body))
-    assert any('"event": "partial"' in chunk for chunk in chunks)
-
-    class _ErrorASR:
-        def transcription(self, _path):
-            return ""
-
-        def stream_transcription(self, _path):
-            raise RuntimeError("stream asr boom")
-
-    _set_request({"stream": "true"}, {"file": _DummyUploadFile("audio.wav")})
-    monkeypatch.setattr(module, "LLMBundle", lambda *_args, **_kwargs: _ErrorASR())
-    monkeypatch.setattr(module.os, "remove", lambda _path: (_ for _ in ()).throw(RuntimeError("cleanup boom")))
-    resp = _run(handler("tenant-1"))
-    chunks = _run(_collect_stream(resp.body))
-    assert any("stream asr boom" in chunk for chunk in chunks)
-
-
-@pytest.mark.p2
-def test_tts_embedded_stream_and_error_matrix_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
-    handler = inspect.unwrap(module.tts)
-    monkeypatch.setattr(module, "get_request_json", lambda: _AwaitableValue({"text": "A。B"}))
-    monkeypatch.setattr(module, "Response", _StubResponse)
-
-    tenant_llm_service = sys.modules["api.db.services.tenant_llm_service"]
-    monkeypatch.setattr(tenant_llm_service.TenantService, "get_by_id", lambda _tid: (False, None))
-    res = _run(handler("tenant-1"))
-    assert res["message"] == "Tenant not found!"
-
-    monkeypatch.setattr(tenant_llm_service.TenantService, "get_by_id", lambda _tid: (True, SimpleNamespace(asr_id="", tts_id="", llm_id="", embd_id="", img2txt_id="", rerank_id="")))
-    res = _run(handler("tenant-1"))
-    assert res["message"] == "No default TTS model is set"
-
-    class _TTSOk:
-        def tts(self, txt):
-            if not txt:
-                return []
-            yield f"chunk-{txt}".encode("utf-8")
-
-    monkeypatch.setattr(tenant_llm_service.TenantService, "get_by_id", lambda _tid: (True, SimpleNamespace(asr_id="", tts_id="tts-x", llm_id="", embd_id="", img2txt_id="", rerank_id="")))
-    monkeypatch.setattr(module, "LLMBundle", lambda *_args, **_kwargs: _TTSOk())
-    resp = _run(handler("tenant-1"))
-    assert resp.mimetype == "audio/mpeg"
-    assert resp.headers.get("Cache-Control") == "no-cache"
-    assert resp.headers.get("Connection") == "keep-alive"
-    assert resp.headers.get("X-Accel-Buffering") == "no"
-    chunks = _run(_collect_stream(resp.body))
-    assert any("chunk-A" in chunk for chunk in chunks)
-    assert any("chunk-B" in chunk for chunk in chunks)
-
-    class _TTSErr:
-        def tts(self, _txt):
-            raise RuntimeError("tts boom")
-
-    monkeypatch.setattr(module, "LLMBundle", lambda *_args, **_kwargs: _TTSErr())
-    resp = _run(handler("tenant-1"))
-    chunks = _run(_collect_stream(resp.body))
-    assert any('"code": 500' in chunk and "**ERROR**: tts boom" in chunk for chunk in chunks)
-
-
-@pytest.mark.p2
 def test_build_reference_chunks_metadata_matrix_unit(monkeypatch):
-    module = _load_session_module(monkeypatch)
+    module = _load_openai_api_module(monkeypatch)
 
     monkeypatch.setattr(module, "chunks_format", lambda _reference: [{"dataset_id": "kb-1", "document_id": "doc-1"}])
     res = module._build_reference_chunks([], include_metadata=False)
